Bethany/requirements.txt
matplotlib==3.9.1
numpy==1.26.4
Pillow==8.3.2
trimesh==3.23.5
count-tokens==0.7.0
opencv-python==4.10.0.84